fix clang warnings
CMakeLists: enable warnings as errrors for clang
library/mysql.parser/mysql-scanner.cpp: fix sign compare warning
library/mysql.parser/CMakeLists.txt: supress parser generated files warnings
backend/wbpublic/CMakeLists.txt: supress parser generated files warnings

+++ backend/wbpublic/grtdb/editor_table.cpp
@@ -383,7 +383,7 @@ bool TableColumnsListBE::set_column_type(const NodeId &node, const GrtObjectRef 
 }
 
 
-IconId TableColumnsListBE::get_field_icon(const NodeId &node, int column, IconSize size)
+IconId TableColumnsListBE::get_field_icon(const NodeId &node, size_t column, IconSize size)
 {
   if (node[0] < real_count())
   {
